requirements.txt
Django==1.6.2
Pillow==2.4.0
South==0.8.4
Unidecode==0.04.14
dj-database-url==0.3.0
dj-static==0.0.5
django-allauth==0.16.1
django-autoslug==1.7.2
django-cities-light==2.4.1
django-taggit==0.12
django-toolbelt==0.0.1
gunicorn==18.0
mysql-python==1.2.5
oauthlib==0.6.1
psycopg2==2.5.3
pystache==0.5.4
python-openid==2.2.5
requests==2.2.1
requests-oauthlib==0.4.0
six==1.6.1
static==1.0.2
wsgiref==0.1.2
yolk==0.4.3
